✨ Skeleton of cuda
