+++ src/cpp/ipc_cuda.cpp
@@ -1,0 +1,6 @@
+#include "ipc_cuda.h"
+
+DLManagedTensor* mtl_tensor_from_cuda_ipc_handle(void *cuda_ipc_handle, int width, int height) {
+    // TODO: Implement this function
+    return nullptr;
+}
